config: use `strictEffects`
This is one of Araq's recommended options [1] for the recent Nim 1.6
release.

To illustrate, recall that `func` is syntactic sugar for the
`noSideEffect`, and consider the below code.

    proc doSomething(n: int) =
      echo n

    func foo(n: int, bar: proc(x: int)) =
      bar(n)
      # Without strictEffects: the above line compiles without error
      # With strictEffects: it produces:
      # /tmp/this_module.nim(4, 6) Error: 'foo' can have side effects
      # > /tmp/this_module.nim(5, 6) Hint: 'foo' calls routine via hidden pointer indirection

    foo(42, doSomething)

Thus, one use of strictEffects is that we can enforce that `foo` is
instead written as a regular (not-func) proc like:

    proc foo(n: int, bar: proc(x: int)) =

or we can annotate the parameter so that `bar` is not allowed to have
side effects:

    func foo(n: int, bar: proc(x: int) {.noSideEffect.}) =

in which case we would see a compile-time error if we used the above
implementation of `doSomething` as the parameter.

See also: the Nim Manual [2]

[1] https://forum.nim-lang.org/t/8404#54227
[2] https://nim-lang.org/docs/manual.html#effect-system