github(publish-release): create only one draft release
In the latest release, commit ecc22452d2d7 caused a new draft release to
be created for each asset - I had to combine these releases manually.
What what we want is to create only one draft release, and for each
build job to upload assets to the same release.

We previously relied on the fact that calling

    gh release create

multiple times creates only one release, but it turns out that calling

    gh release create --draft

multiple times creates multiple draft releases.

With this commit, we run the latter command only if there is no release
yet for the tag.