sync: fix `test_runner` field type
First, the previous `string` should have at least been `bool`. Oops.

But secondly, for a missing `test_runner` key, `jsony` by default sets
the corresponding field value to `false`. However, the spec says that an
omitted `test_runner` key implies the value `true`, and so we need to
distinguish between the "missing key" case and the "key is present and
set to false" case.

Let's just use `Option[bool]` to do this.

Every other key is a string or an array of strings - if a
`.meta/config.json` file specifies the value of one of the other
optional keys as the empty array or empty string, we can just omit it
when we write the updated JSON.

That is, we don't need `Option` for the other optional keys, unless we
want to support JSON `null` for one of these keys. For example:

    "contributors": null,

currently produces a JSON parsing error. Right now, `configlet sync
--metadata` works on every track except the `common-lisp` track, because
it contains such a `null`.

Instead, we might want to just remove that `null` when writing the
updated JSON.