sync_common, sync_filepaths, sync_metadata: reduce string allocations
This is a worthwhile optimization, even if just so it doesn't distract
me when profiling.

Before this commit, when running an offline

    $ configlet sync --filepaths

and

    $ configlet sync --metadata

about 12.5% and 20% of the respective execution times were spent in
`joinPath`, just constructing the paths to exercise `.meta/config.json`
files (and `tests.toml` files in the latter case). That's pretty
expensive for what it is.

This commit removes that `joinPath` overhead. Now about 95% of the time
is spent parsing JSON or TOML files.