sync_common: fix `test_runner` serialization
Before this commit, a `test_runner` value was incorrectly serialized as
a string:

    "test_runner": "false"

With this commit, it is serialized as a boolean:

    "test_runner": false

This bug happened because I was previously using `escapeJsonUnquoted`
everywhere and adding the quotes manually - passing "false" to one of
the procs could indeed omit the quotes.

But now we use `escapeJson` everywhere, which adds the quotes. So let's
add a separate `addBool` procedure.