test_sync: test serialization against round-trip
This checks that we serialize every exercise `.meta/config.json` file in
the `exercism/elixir` repo the same as a somewhat-convoluted round-trip
serialization process that uses `jsony.toJson` -> `json.parseJson` and
then removes certain key/value pairs.

A problem with this test is that the round-trip serialization might be
more bug-prone and harder to reason about anyway - if the test
fails, the problem might be in the round-trip serialization.

For one thing, it's easy to miss or introduce a bug like

    if j["icon"].len == 0:
      delete(j, "icon")

which compiles, but doesn't do what was intended. Instead:

    if j["icon"].getStr().len == 0:
      delete(j, "icon")