build: pin dependency versions
The version of Nimble that's distributed with Nim does not yet support
lock files. But until it does, we can use `==` as a lesser step in the
right direction.

Unfortunately, a person with write access to these dependencies could
still theoretically tag a different (potentially bad/malicious) commit
with the tag we specify here.

Note that this commit does not pin the Nim version in this file, so that
we support a configlet developer who uses the `devel` Nim
compiler locally (currently at version 1.7.1). However, the configlet
release pipeline _does_ pin the Nim version (at 1.6.0).