nimble: move `isaac` below `uuids`
To actually use the pinned `isaac` version, it turns out that we must
specify `isaac-#45a5cbbd54ff59ba3ed94242620c818b9aad1b5b` _after_
`uuids`, which has `isaac` as a dependency.

This is necessary so that when running

    $ nimble build --verbose

the path to `isaac-#45a5cbbd54ff59ba3ed94242620c818b9aad1b5b` comes at
the end, and therefore takes precedence:

    Executing /usr/bin/nim c \
      --colors:on \
      --noNimblePath \
      -d:NimblePkgVersion=4.0.0 \
      --path:'/home/foo/.nimble/pkgs/cligen-#b962cf8bc0be847cbc1b4f77952775de765e9689' \
      --path:'/home/foo/.nimble/pkgs/jsony-#eb63a326b7f16537764c090f8859eb2451ad8d4d' \
      --path:'/home/foo/.nimble/pkgs/parsetoml-#9cdeb3f65fd10302da157db8a8bac5c42f055249' \
      --path:'/home/foo/.nimble/pkgs/uuids-#8cb8720b567c6bcb261bd1c0f7491bdb5209ad06' \
      --path:/home/foo/.nimble/pkgs/isaac-0.1.3 \
      --path:'/home/foo/.nimble/pkgs/isaac-#45a5cbbd54ff59ba3ed94242620c818b9aad1b5b'  \
      -o:/home/foo/configlet/configlet \
      /home/foo/configlet/src/configlet.nim

Before this commit, `nimble build` uses the latest `isaac` tag,
even if that does not correspond to the hash in the .nimble file (!).