cli, tests: replace `--mode choose|include|exclude`
`configlet sync` previously operated only on tests, and had the `--mode`
option to control how missing tests were handled. But `configlet sync`
now also operates on docs, filepaths, and metadata, and these don't have
the same "include, exclude, or skip?" decision.

With this commit, we instead use

    --tests choose|include|exclude

with a default value of `choose`.

The alternatives seem worse. In particular, with

    --tests --mode include

the user is likely to wonder "why does something called --mode just
apply to --tests?"

And the alternative of

    --tests --tests-mode include

seems overly verbose, by the argument: "if --tests-mode just applies
to --tests, and there is no other option to customize how --tests works,
can't --tests just take the --tests-mode value?"