cli: improve `--filepaths` help description
`--filepaths` is an option for the `sync` command, but it isn't
"syncing" in the way that the other options do.

The other options keep the track data up to date when it exists in
problem-specifications, but filepaths data comes from the track's own
`config.json` file. This commit clarifies that

    $ configlet sync --filepaths

will preserve any exercise config `files` value that is already
populated.