build: patch parseopt3
Before this commit, we could only write `--tests` without a value if we
wrote it at the end of the command line. For example, running

    $ configlet sync --tests -e yacht

would produce

    Error: invalid value for '--tests': '-e'

This is because:

- The command line parser knows which options can take a value, and
  supports a value starting with the `-` character. That is, it does not
  naively just assume that every parameter starting with `-` is an
  option.

- The `--tests` option is special amongst our options. It can take a
  value of `choose|include|exclude`, because a separate `--tests-mode`
  option seems overly verbose. But it makes sense for just `--tests`
  alone to work, and do the same as `--tests choose`, because `--docs`,
  `--filepaths` and `--metadata` all work (these options do not take a
  value).

As a workaround, this commit patches the command line parsing so that if
we have a long option followed by a space and a parameter that begins
with the `-` character, that parameter is always parsed as an option,
not a value.

In this commit, we decide whether to patch based on the parsing result.
But this has a problem: it can only work when the parsing library is
already installed. See the next commit.