build: make patch work in CI
The previous commit was an interesting implementation, but doesn't work
unless cligen is already installed (because it imports it). So it
doesn't work in CI as-is during the initial setup. We could amend the
build script such that it installs cligen first, but that means a plain
`git clone` of the configlet repo followed by `nimble build` won't work.

So let's just decide whether to patch based on the hash of
`parseopt3.nim`, rather than actually testing that something parses as
we want. This means we'll have to update the hash if `parseopt3.nim`
gets updated upstream. But we'll soon pin every Nimble package.

Here, I've just used `std/hashes` because the `std/sha1` (or `std/md5`)
module isn't available in the Nim VM, and calling an external process
via e.g. `sha512sum parseopt3.nim` is less portable.

We could instead just add the entire patvched `parseopt3.nim` file to
our repo, as we did we the patched `std/json` and `std/parsejson`. But
let's try it this way for now, arguing that a patch is more
self-documenting than a forked file.