sync_common, test_sync: fix serialization of `custom: {}`
Before this commit, if we ran

    $ configlet sync --update

and an exercise `.meta/config.json` file had:

- unsynced metadata, or a sync-able `files` value
- and a `custom` key with the value of the empty object

for example:

    {
      ...
      "blurb": "some unsynced blurb",
      "custom": {}
    }

because I hadn't got around to handling this case yet, configlet would
write invalid JSON:

    {
      ...
      "blurb": "some now-synced blurb",
      "custom":
    }

With this commit, we instead omit the `custom` key/value pair when its
value is the empty object, instead producing:

    {
      ...
      "blurb": "some now-synced blurb"
    }

which is consistent with the behavior for other optional key/value pairs
that are empty (e.g. `"contributors": []`).

If the value of the `custom` key is not a JSON object,
`configlet sync --update` currently complains at serialization-time and
quits. I'd prefer to complain at parse-time, so we can provide filename
and line number information.