sync_common, test_sync: work towards preserving key order
This does a second pass over the JSON to get the key order. Using
`getTok` directly is a little faster than using `json.parseFile`.

Ideally, we'd have only one (jsony) pass. However, I didn't see a clean
way to do that until a later commit.

I'm pushing this commit so that the reader can see what a second pass
would entail, and how it's nicer to avoid it. You don't need to review
this `getJsonKeys` implementation.