sync_common, test_sync: use jsony `renameHook` instead
This is one of the more interesting commits on this branch, and it shows
the strengths of jsony's hookable design.

With this commit, we record the original key order as we do a single
parse of the JSON, even though the parsing library tries not compromise
on speed, and therefore:
- does not keep track of the key order
- and directly populates a strongly typed object (whose fields are in a
  fixed order) with no intermediate allocations

I read the jsony source code again, and found that during its object
parsing, `renameHook` has exactly the context that we want. So not only
do we avoid a relatively expensive second pass, which more than doubled
the JSON parsing time (and made it so that it'd be faster to do a single
pass into `JsonNode` and then use the `keys` iterator), we do so without
redefining the whole `parseHook`.

This also gives us a cleaner way to separate the key order of the
`files` keys.

As of this commit, we're still writing the "formatted" JSON as the
upcoming `configlet fmt` will write it in the future. But the
serialization proc now has access to the original key order, and so a
later commit can make it so that `configlet sync` preserves the key
order.