sync_common: temporarily remove more `fmt` code
This commit guarantees that we don't accidentally call `pretty` in
"format mode" somewhere.

Let's add this code back in the commit that adds `configlet fmt`
instead.

The removed code also had a minor bug, which will be fixed in that
upcoming commit: `fmt` shouldn't write the `editor` key when its value
is the empty array.