cli: don't silently accept `format`
Silently allowing `configlet format` might be unwise - somebody might
use it in a script and rely on it being available, and it could cause
confusion if both `fmt` and `format` appear in the wild.

So let's omit this feature for now. If we want to add it later, it's a
backward-compatible change. But if we add it, then remove it later, it
could break somebody's script.

There's an argument that minimizing undocumented "conveniences" and
features is a good idea. A recent post that's somewhat related:
https://daniel.haxx.se/blog/2021/12/06/no-easter-eggs-in-curl/