fmt: add initial implementation (#468)
This commit adds the `configlet fmt` command, to handle the rewriting of
JSON files on a track in a canonical form. Currently, it operates only
on exercise `.meta/config.json` files - later, it will likely also
format the other JSON files on a track:

- The track `config.json` file.

- Concept `.meta/config.json` files and `links.json` files.

The interface is similar to `configlet sync`: a plain `configlet fmt`
makes no changes to the track, and checks the `.meta/config.json` file
formatting for every Concept Exercise and Practice Exercise.
Furthermore, `fmt` supports the same `-e/--exercise`, `-u/--update`,
and `-y/--yes` options that `sync` uses.

This design makes it easy to use `configlet fmt` in CI. A similar `fmt`
command existed in the configlet for Exercism v2, but it had no dry-run
ability: to use in CI, we needed to run `git diff --exit-code` or
similar afterwards.

For more details on the new `fmt` functionality and usage examples,
please see the changes to the `README.md` file.

The implemented canonical key ordering for an exercise
`.meta/config.json` file is:

    - authors
    - [contributors]
    - files
      - solution
      - test
      - exemplar           (Concept Exercises only)
      - example            (Practice Exercises only)
      - [editor]
    - [language_versions]
    - [forked_from]        (Concept Exercises only)
    - [icon]               (Concept Exercises only)
    - [test_runner]        (Practice Exercises only)
    - blurb
    - [source]
    - [source_url]
    - [custom]

where [key] indicates that the key is optional. This order tries to:

- Maximize visual consistency between Concept Exercise and
  Practice Exercise configs by writing their unique keys in the same
  place.

- Maximize visual consistency between configs that contain optional keys
  and those that lack them, by writing those keys later (e.g. `custom`).
  The exceptions are:

  - Write `contributors` under `authors`, since these keys are similar.

  - Write `blurb` near the bottom, and above `source`, and `source_url`,
    since these keys are all synced with `problem-specifications` for a
    Practice Exercise.