cli, README: add command descriptions to help message (#469)
This commit makes the output of `configlet --help` more useful.

Before this commit, we listed only the name of each available command:

    $ configlet --help
    Usage:
      configlet [global-options] <command> [command-options]

    Commands:
      fmt, lint, sync, uuid, generate, info

    [...]

With this commit, we also have a short description of each command:

    $ configlet --help
    Usage:
      configlet [global-options] <command> [command-options]

    Commands:
      fmt       Format the exercise 'meta/config.json' files
      generate  Generate Concept Exercise 'introduction.md' files from 'introduction.md.tpl' files
      info      Print some information about the track
      lint      Check the track configuration for correctness
      sync      Check or update Practice Exercise docs, metadata, and tests from 'problem-specifications'.
                Check or populate missing 'files' values for Concept/Practice Exercises from the track 'config.json'.
      uuid      Output new (version 4) UUIDs, suitable for the value of a 'uuid' key

    [...]

This is especially useful after the previous two commits, which expanded
the `sync` command and added the `fmt` command.

We also now order the commands alphabetically.

In the future, configlet may gain per-command help messages. For
example, `configlet sync --help` could produce a longer description of
the `sync` command and usage examples.