info: omit deprecated exercises in unimplemented exercises list (#474)
When running `configlet info`, the list of unimplemented exercises on a
track previously included exercises that are deprecated in
`problem-specifications`. With this commit, we show only non-deprecated
exercises.

Example output before this commit:

    $ configlet info

    [...]

    There are 11 exercises from `exercism/problem-specifications` that are neither implemented nor in the track config `exercises.foregone` array:
    counter
    error-handling
    hangman
    ledger
    lens-person
    nucleotide-codons
    octal
    paasio
    point-mutations
    tree-building
    trinary

    [...]

With this commit:

    $ configlet info

    [...]

    There are 6 non-deprecated exercises in `exercism/problem-specifications` that
    are both unimplemented and not in the track config `exercises.foregone` array:
    error-handling
    hangman
    ledger
    lens-person
    paasio
    tree-building

    [...]

For now, we do this by parsing a simple custom file format. In the
future, we could stop baking the slugs into the binary: `configlet info`
could use a (cached) `problem-specifications` directory at run-time,
similar to `configlet sync`.

Closes: #473