lint(track_config): speedup cycle checking slightly
We were using concatenation of two sequences [1]:

    proc `&`*[T](x, y: seq[T]): seq[T] {.noSideEffect.} =
      newSeq(result, x.len + y.len)
      for i in 0..x.len-1:
        result[i] = x[i]
      for i in 0..y.len-1:
        result[i+x.len] = y[i]

but we can use the `&` overload that appends one element [2]:

    proc `&`*[T](x: seq[T], y: T): seq[T] {.noSideEffect.} =
      newSeq(result, x.len + 1)
      for i in 0..x.len-1:
        result[i] = x[i]
      result[x.len] = y

The Nim compiler doesn't optimize the creation+appending of a
one-element sequence into an append of the element. Somewhat similarly,
the compiler deliberately doesn't optimize `string1 = string1 & string2`
- it should be written as `string1.add string2` or `string1 &= string2`.

The concept cycle checking is a hotspot on tracks with a larger number
of Concept Exercises - for example, with the csharp track, about a third
of the `configlet lint` execution time was spent in `checkForCycle`.
This commit removes some unnecessary allocations within that hotspot,
and thereby produces a small (but measureable) `configlet lint` speedup
of 3-5% on such tracks. For the csharp track:

    $ hyperfine -i \
      --warmup 100 \
      --runs 100 \
      -L binary configlet_before,configlet_after \
      './{binary} lint' \
      --export-markdown /tmp/bench_configlet_lint_csharp.md

| Command                   |  Mean [ms] | Min [ms] | Max [ms] |    Relative |
|:--------------------------|-----------:|---------:|---------:|------------:|
| `./configlet_before lint` | 39.9 ± 0.3 |     39.2 |     41.1 | 1.04 ± 0.01 |
| `./configlet_after lint`  | 38.5 ± 0.4 |     38.0 |     40.4 | 1.00        |

[1] https://github.com/nim-lang/Nim/blob/b812431f8317/lib/system.nim#L1672-L1686
[2] https://github.com/nim-lang/Nim/blob/b812431f8317/lib/system.nim#L1688-L1701