config: disable `strictEffects`
Before this commit, `nimble build` produced the following error:

    /home/runner/.nimble/pkgs/parsetoml-#9cdeb3f65fd10302da157db8a8bac5c42f055249/parsetoml.nim(1745, 27) template/generic instantiation of `==` from here
    /home/runner/.choosenim/toolchains/nim-1.6.2/lib/system/comparisons.nim(284, 6) Error: '==' can have side effects
    > /home/runner/.choosenim/toolchains/nim-1.6.2/lib/system/comparisons.nim(309, 13) Hint: '==' calls `.sideEffect` '=='
    >> /home/runner/.nimble/pkgs/parsetoml-#9cdeb3f65fd10302da157db8a8bac5c42f055249/parsetoml.nim(1725, 6) Hint: '==' called by '=='

We can't disable `strictEffects` on a push/pop or per-file basis, and it
seems like we can't resolve the problem with a one-line patch/PR in
`parsetoml`. So let's just disable `strictEffects` for now.

The upstream plan is for `strictEffects` to become the default behavior
in the future, so I suspect this issue will resolve itself.