info: refactor to use JSON prob-specs slugs list (#487)
The current implementation of `configlet info` requires the list of
`problem-specifications` exercise slugs to be baked into the configlet
executable. We did that simply by writing the slugs newline-separated in
a file, and parsing that into a `HashSet` at compile-time. This is
cleaner than writing every slug in a `const` in a nim file if we wanted
to update the list automatically.

Then, we wanted `configlet info` to omit upstream-deprecated exercises
when listing the unimplemented exercises on a track. So it made sense to
group the slugs by kind (rather than just omitting or commenting-out the
deprecated slugs), and commit f798115c2250 did that via a simple custom
file format and parser.

With this commit, we refactor to use a JSON-formatted list of prob-specs
slugs, and thereby remove the custom parsing code. This also makes it
easier to bake-in more information that configlet could use, like a
timestamp and commit ref for the prob-specs state.

We still parse the file and populate `HashSet`s at compile-time, so the
main cost is the more verbose JSON syntax.

In the future, `configlet info` will probably use a cached clone of
`problem-specifications`, and respect the `--offline` and
`--prob-specs-dir` options that work for `configlet sync`. Then we won't
need to include the slugs in the configlet executable, nor update that
list whenever an exercise in prob-specs is added, deprecated, or gains
canonical data.

This is a refactoring-only commit. It doesn't change the list of
exercises in `problem-specifications`.

Closes: #482