cli: set help text at compile time
Before this commit, the name of the executable in the help message was
set at execution time. That is, renaming the configlet executable to
`foo` and then running `foo --help` would produce a help message that
begins:

    Usage:
      foo [global-options] <command> [command-options]

We don't need to support this, so let's set the name of the application
at compile time. This saves a call to `getAppFilename`, and makes it
cleaner to add a short description of configlet to the top of the help
message in the future.