sync: start work on caching prob-specs repo
Before this commit, `configlet sync` would make a temporary clone of
problem-specifications when the `--prob-specs-dir` option was not given.

With this commit, it now looks for a cached copy of
problem-specifications on the user's machine.

On Linux, the cached path is:

    ~/.cache/exercism/configlet/problem-specifications

This commit works, but it doesn't yet handle the upstream prob-specs
being updated - it produces an error saying the local prob-specs is not
up to date (which is also what happens when a directory given by
`--prob-specs-dir` is not up to date).

We should automatically `git pull` in the cached dir when reasonable.