cli: refer to next release as `4.0.0.beta.1`
Configlet has been relatively stable and used in production for quite
some time. We avoided releasing the stable `4.0.0` because there were a
few breaking changes we wanted to make first, but I've suggested moving
to `beta` for the remainder of the development phase.

This helps to convey that configlet is getting closer to `4.0.0` stable,
and probably won't set your machine on fire. Declaring `beta` at this
exact moment also helps us to distinguish configlet versions that have
the new caching behavior.