sync: cache prob-specs repo, and remove `--prob-specs-dir` option (#540)
Before this commit, the default behavior of `configlet sync` was to
clone `exercism/problem-specifications` and delete it afterwards. The
`-p, --prob-specs-dir` and `-o, --offline` options influenced the
behavior as shown in the below table, where an `x` in one of the first
two columns means the corresponding option was given.

| -p | -o | Behavior                                              |
| -- | -- | ----------------------------------------------------- |
|    |    | Clone prob-specs, and delete it afterwards            |
| x  |    | Use the given prob-specs dir; error if outdated       |
| x  | x  | Use the given prob-specs dir; don't check if outdated |
|    | x  | Error: `-p` must be provided when `-o` is given       |

This was simple and robust, but slow in the zero-option case. Most track
maintainers would probably want to use both `--offline` and
`--prob-specs-dir` nearly all the time, especially because we often want
to run several `configlet sync` commands in quick succession.

With this commit, `configlet sync` now looks for a cached
problem-specifications on the user's machine, behaving as follows:

| Cache hit? | -o | Behavior                                           |
| ---------- | -- | -------------------------------------------------- |
|            |    | Clone prob-specs, and don't delete it afterwards   |
|     x      |    | Use cached prob-specs, and `git pull` in it        |
|     x      | x  | Use cached prob-specs, and don't `git pull` in it  |
|            | x  | Error: local prob-specs not found, and won't clone |

This optimizes for the convenience of `configlet sync` and
`configlet sync --offline` - to perform an offline sync, the user no
longer needs to also write a path to a prob-specs directory.

In fact, this commit also removes the `--prob-specs-dir` option
entirely; its remaining usefulness is not worth the added complexity in
the help message (or the code).

Despite the `--prob-specs-dir` removal, note that you can still:

- Sync from an older prob-specs commit, by checking out that commit in
  the cached repo and then using `--offline`. We require the commit
  to exist on the upstream `main` branch.

- Sync using `--offline` when your only copy of problem-specifications
  is outside the cache directory (by copying to the cached location).

The path to the cached prob-specs is:

    $PATHSTART/exercism/configlet/problem-specifications

where the value of `$PATHSTART` is given by this table:

| Platform | $PATHSTART                                                  |
| -------- | ----------------------------------------------------------- |
| Windows  | getEnv("LOCALAPPDATA")                                      |
| macOS    | getEnv("XDG_CACHE_HOME", getEnv("HOME") / "Library/Caches") |
| Other    | getEnv("XDG_CACHE_HOME", getEnv("HOME") / ".cache")         |

and `getEnv` returns the value of the environment variable with the name
of the first parameter (or if that does not exist, the value of the
second parameter).

For example, on Linux, the cached path is typically:

    ~/.cache/exercism/configlet/problem-specifications

Another benefit is that other configlet commands can use the cached
`problem-specifications` too. For example, `configlet info` currently
uses some prob-specs data that is baked into the configlet binary; we
can remove that data in the future.

Closes: #477