release: 4.0.0-beta.1
Configlet has been relatively stable and used in production for quite
some time. We avoided releasing the stable `4.0.0` because there were a
few breaking changes we wanted to make first, but I've suggested moving
to `beta` now that we're closer to the end of this development phase.

This helps to convey that configlet is not "dangerous alpha software".
Also, declaring `beta` at this exact moment helps us to distinguish
configlet versions that have the new caching behavior (and lack the
--prob-specs-dir option).