build(publish-release): upload first asset at release creation (#550)
It seems that the first build job to finish (usually the Linux one)
would sometimes fail in the `gh release upload` step, saying that the
release was not found. To address this race condition, upload the first
asset during the release creation.

From the docs for `gh release create` [1]:

    gh release create [<tag>] [<files>...]

    Create a new GitHub Release for a repository.

    A list of asset files may be given to upload to the new release.

Note that we still have another race condition: a separate job can
create a release in the time between `gh release view` indicating
"there is no such release" and `gh release create` running. We should
address that later, but it hasn't caused a problem yet - the Linux job
consistently finishes significantly faster than the others. And if we
encounter such a failure in the meantime, the workaround is simple: just
restart the failing job.

[1] https://cli.github.com/manual/gh_release_create