info: add script for writing prob-specs JSON file (#555)
Running this script will update `src/info/prob_specs_exercises.json`
locally. I ran this script to produce the diff in the previous commit,
b54fe48f7290.

We will no longer need this script if we alter `configlet info` to use
the new prob-specs cache, rather than baked-in prob-specs data. But it's
worth having the script in the git history, even if we delete it later.

For background, see commit b8432f601490 (which added the initial JSON
file, created by an older version of this script).