release: 4.0.0-beta.1 (#547)
In this repo we have developed the configlet for Exercism v3, which is
a complete rewrite of the configlet for Exercism v1/v2. This configlet
has stayed "alpha" since January 2021 - partly for the freedom to make
breaking changes while we converge on a CLI design, but partly from
inertia. In reality, configlet has been working reliably for a long
time; the latest "alpha release" runs in CI on every PR on every track,
and is downloaded whenever a user runs a `fetch-configlet` script.

We're getting closer to releasing 4.0.0, but let's start `beta` releases
in the meantime - this better conveys the current situation, and that
configlet is not "dangerous alpha software".

Declaring `beta` at this exact moment also helps us to distinguish
configlet versions that have the new caching behavior (and lack the
`--prob-specs-dir` option).

Refs: #524