rearrange some procedures
Previous commits in this series moved some types to a common file. This
meant that, for example, `info.nim` no longer needed to import
`lint/track_config` to import the types. However, it still needed to
import it for the `$` and the `init` procedures.

This commit improves a little, moving the minimum procedures to:
- Deduplicate the `$` funcs that existed in both `lint/track_config`
  and `sync/sync_common`.
- Allow `info.nim` to no longer import `lint/track_config`.

The name "types_track_config" is perhaps misleading now, however (since
it's not _just_ types). But I think it's an improvement overall.