sync_common: fix comment
`configlet sync` does not reorder keys, unlike `configlet fmt`. So if
the `editor` key is present, but is not the final key, a missing
`invalidator` key is not placed directly after it - with the current
code, it is always placed at the end.

We `insert` the other missing keys at a certain position, based on the
previous keys (which are all required). However, there may not be an
`editor` key. Instead of conditionally inserting after `editor`, let's
just keep the current code and make the comment match it.