github(workflows): bump GCC from 9.4 to 10.3
The Linux jobs currently use `ubuntu-20.04`, which currently uses
GCC 9.4 by default. However, GCC 10.3 is also already installed [1], and
we can use via `update-alternatives` [2]. It is not sufficient to only
set the `CC` environment variable.

At the time of writing, the `ubuntu-22.04` label is available as a
public beta, and it adds GCC 11.2 [3]. The latest release is GCC 12.1.

Since GCC 5.1 (2015), GCC receives a major version bump once per year
(in late April or early May). Before that, "the GCC major number carried
little to no useful information" [4].

Here is an abridged history of recent GCC releases [5]:

    Version  Date
    -------  ----------
       12.1  2022-05-06
       11.3  2022-04-21
       11.2  2021-07-28
        9.4  2021-06-01
       11.1  2021-04-27
       10.3  2021-04-08
       10.2  2020-07-23
       10.1  2020-05-07
        9.3  2020-03-12
        9.2  2019-08-12
        9.1  2019-05-03
        8.1  2018-05-02
        7.1  2017-05-02
        6.1  2016-04-27
        5.1  2015-04-22

With the configlet 4.0.0-beta.2 Linux release binary:

    $ readelf -p .comment configlet

    String dump of section '.comment':
      [     0]  GCC: (Ubuntu 9.2.1-12ubuntu1) 9.2.1 20191022
      [    2d]  GCC: (Ubuntu 9.4.0-1ubuntu1~20.04.1) 9.4.0

With a configlet built in the GitHub Actions environment using this
commit:

    $ readelf -p .comment configlet

    String dump of section '.comment':
      [     0]  GCC: (Ubuntu 9.2.1-12ubuntu1) 9.2.1 20191022
      [    2d]  GCC: (Ubuntu 10.3.0-1ubuntu1~20.04) 10.3.0

[1] https://github.com/actions/virtual-environments/blob/5a2cb18a48bc/images/linux/Ubuntu2004-Readme.md
[2] https://manpages.ubuntu.com/manpages/focal/en/man1/update-alternatives.1.html
[3] https://github.com/actions/virtual-environments/blob/5a2cb18a48bc/images/linux/Ubuntu2204-Readme.md
[4] https://gcc.gnu.org/develop.html
[5] https://gcc.gnu.org/releases.html

Closes: 193